only one version V0 left. The 'legacy' version is now the current KafkaMappedConnector in the API
