+++ obp-ri-demo/src/main/resources/logback.xml
@@ -31,7 +31,7 @@
 
   <logger name="com.tesobe.obp" level="trace"/>
   <logger name="com.tesobe.obp.transport.spi.Connector" level="trace"/>
-  <logger name="com.tesobe.obp.transport.spi.ResponderV1" level="trace"/>
+  <logger name="com.tesobe.obp.transport.spi.ResponderV0" level="trace"/>
 
   <logger name="com.tesobe.obp.transport.spi.LoggingReceiver" level="info"
     additivity="false">
