+++ obp-ri-transport/src/test/resources/logback-test.xml
@@ -26,7 +26,7 @@
 
   <logger name="com.tesobe.obp" level="debug"/>
   <logger name="com.tesobe.obp.transport.spi.Connector" level="trace"/>
-  <logger name="com.tesobe.obp.transport.spi.ResponderV1" level="trace"/>
+  <logger name="com.tesobe.obp.transport.spi.ResponderV0" level="trace"/>
 
   <logger name="com.tesobe.obp.transport.spi.SpeedTest" level="info"
     additivity="false">
